+++ packages/app-degree-pages/.storybook/main.js
@@ -1,5 +1,4 @@
 const path = require("path");
-const PROJECT_DIR = path.resolve(__dirname, "../");
 const common = require("../webpack/webpack.common");
 
 module.exports = {