+++ packages/app-degree-pages/src/components/DetailPage/index.js
@@ -3,6 +3,8 @@ import { Hero } from "@asu/components-core";
 import PropTypes, { arrayOf } from "prop-types";
 import React, { useContext, useEffect, useState } from "react";
 
+// @ts-ignore
+import trackReactComponent from "../../../../../shared/services/componentDatalayer";
 import { useFetch } from "../../../../../shared";
 import {
   ErrorAlert,
@@ -87,6 +89,19 @@ const DetailPage = ({
   const url = urlResolver(dataSource, detailPageDefaultDataSource);
   const { defaultState } = useContext(AppContext);
   const { detailPageDefault } = defaultState;
+
+  useEffect(() => {
+    if (typeof window !== "undefined") {
+      trackReactComponent({
+        packageName: "app-degree-pages",
+        component: "DetailPage",
+        type: "NA",
+        configuration: {
+          dataSource,
+        },
+      });
+    }
+  }, []);
 
   useEffect(() => {
     doFetchPrograms(url);