+++ packages/component-cookie-consent/.storybook/main.js
@@ -1,4 +1,9 @@
 const common = require("../webpack/webpack.common");
+const path = require("path");
+
+const PROJECT_DIR = path.resolve(__dirname, "../");
+
+
 
 module.exports = {
   addons: [